Refactor: Move network classes from util to remote
