+++ firebase-firestore/src/main/java/com/google/firebase/firestore/remote/FirestoreChannel.java
@@ -1,4 +1,4 @@
-// Copyright 2018 Google LLC
+// Copyright 2019 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.firebase.firestore.util;
+package com.google.firebase.firestore.remote;
 
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.TaskCompletionSource;
@@ -21,7 +21,9 @@ import com.google.firebase.firestore.FirebaseFirestoreException;
 import com.google.firebase.firestore.FirebaseFirestoreException.Code;
 import com.google.firebase.firestore.auth.CredentialsProvider;
 import com.google.firebase.firestore.model.DatabaseId;
-import com.google.firebase.firestore.remote.FirestoreCallCredentials;
+import com.google.firebase.firestore.util.AsyncQueue;
+import com.google.firebase.firestore.util.Logger;
+import com.google.firebase.firestore.util.Util;
 import com.google.firestore.v1.FirestoreGrpc;
 import com.google.firestore.v1.FirestoreGrpc.FirestoreStub;
 import io.grpc.CallOptions;
@@ -38,7 +40,7 @@ import java.util.concurrent.TimeUnit;
  * Wrapper class around io.grpc.Channel that adds headers, exception handling and simplifies
  * invoking RPCs.
  */
-public class FirestoreChannel {
+class FirestoreChannel {
 
   private static final Metadata.Key<String> X_GOOG_API_CLIENT_HEADER =
       Metadata.Key.of("x-goog-api-client", Metadata.ASCII_STRING_MARSHALLER);
@@ -66,11 +68,11 @@ public class FirestoreChannel {
   /** The value to use as resource prefix header. */
   private final String resourcePrefixValue;
 
-  public FirestoreChannel(
-      AsyncQueue asyncQueue,
-      CredentialsProvider credentialsProvider,
-      ManagedChannel grpcChannel,
-      DatabaseId databaseId) {
+  FirestoreChannel(
+          AsyncQueue asyncQueue,
+          CredentialsProvider credentialsProvider,
+          ManagedChannel grpcChannel,
+          DatabaseId databaseId) {
     this.asyncQueue = asyncQueue;
     this.credentialsProvider = credentialsProvider;
