+++ firebase-firestore/src/main/java/com/google/firebase/firestore/remote/IncomingStreamObserver.java
@@ -1,4 +1,4 @@
-// Copyright 2018 Google LLC
+// Copyright 2019 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.firebase.firestore.util;
+package com.google.firebase.firestore.remote;
 
 import io.grpc.Metadata;
 import io.grpc.Status;
 
 /** Interface used for incoming/receiving gRPC streams. */
-public interface IncomingStreamObserver<RespT> {
+interface IncomingStreamObserver<RespT> {
   /** Headers were received for this stream. */
   void onHeaders(Metadata headers);
